Reformat vlc autotools plugin to avoid duplication
Inherit from snapcraft autotools plugins and call parent build() one.
Fixes #118

Note that there is a hack in it, as said in the file:
The plugins hooks are not idemnpotent, where they should be.
so we need to answer that calling the autotools plugins won't
retrigger BasePlugin build() which erase the directory.
However the issue with this hack is that other parts from this
project will be impacted if they are instantiated after this
method is ran, which is unlikely, but still possible.
https://bugs.launchpad.net/snapcraft/+bug/1595964.